refactor: fix wrong component deps
