+++ apps/cowswap-frontend/src/modules/bridge/pure/SwapStopDetails/index.tsx
@@ -15,12 +15,12 @@ import { ConfirmDetailsItem } from 'modules/trade/pure/ConfirmDetailsItem'
 import { ReceiveAmountTitle } from 'modules/trade/pure/ReceiveAmountTitle'
 import { UsdAmountInfo } from 'modules/usdAmount/hooks/useUsdAmount'
 
+import { StyledRefundCompleteIcon } from './styled'
+
 import { ArrowIcon, Link, StyledSpinnerIcon, TokenFlowContainer } from '../../styles'
 import { BridgeProtocolConfig, BridgeFeeType } from '../../types'
-import { getFeeTextColor } from '../../utils/fees'
-import { StopStatusEnum } from '../../utils/status'
+import { getFeeTextColor, StopStatusEnum } from '../../utils'
 import { BridgeDetailsContainer } from '../BridgeDetailsContainer'
-import { StyledRefundCompleteIcon } from '../BridgeStopDetails/index'
 import { RecipientWrapper } from '../BridgeStopDetails/styled'
 import { NetworkLogo } from '../NetworkLogo'
 import { TokenAmountDisplay } from '../TokenAmountDisplay'
